% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{binplot}
\alias{binplot}
\title{Visualize binary indicator matrix optionally coupled with cross-validation performance}
\usage{
binplot(
  fit,
  cv,
  kmax,
  collines = TRUE,
  rowlines = TRUE,
  cex.axis = 0.6,
  heights = c(0.2, 0.8),
  ...
)
}
\description{
Visualize binary indicator matrix optionally coupled with cross-validation performance
}
