% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{bs.boxplot}
\alias{bs.boxplot}
\title{Bootstrap visualization with boxplot, percentage of new additions}
\usage{
bs.boxplot(bs, ...)
}
\description{
Bootstrap visualization with boxplot, percentage of new additions
}
