% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bs.casso}
\alias{bs.casso}
\title{Bootstrapping for casso-fitted model objects}
\usage{
bs.casso(fit, bootstrap = 100, seed = NULL, verb = 0, ...)
}
\arguments{
\item{fit}{PARAM_DESCRIPTION}

\item{bootstrap}{PARAM_DESCRIPTION, Default: 100}

\item{seed}{PARAM_DESCRIPTION, Default: NULL}

\item{verb}{PARAM_DESCRIPTION, Default: 0}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[casso]{character(0)}}
}
