% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bs.k}
\alias{bs.k}
\title{FUNCTION_TITLE}
\usage{
bs.k(bs)
}
\arguments{
\item{bs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
