% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{bs.plot}
\alias{bs.plot}
\title{Bootstrap + cross-validation heatmap plot}
\usage{
bs.plot(
  fit,
  bs,
  kmax,
  cex.axis = 0.6,
  palet = colorRampPalette(c("orange", "red", "black", "blue", "cyan"))(dim(bs)[3]),
  nbins = 100,
  Colv = NA,
  Rowv = NA,
  ...
)
}
\description{
Bootstrap + cross-validation heatmap plot
}
