% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{bs.visu}
\alias{bs.visu}
\title{Visualize bootstrapping of a fit casso object}
\usage{
bs.visu(bs, intercept = FALSE)
}
\arguments{
\item{bs}{PARAM_DESCRIPTION}

\item{intercept}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
