% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitS4.R
\name{casso}
\alias{casso}
\title{FUNCTION_TITLE}
\usage{
casso(x, y, k, w, family = "gaussian", print = 3, start = 2, verb = 1, kmax)
}
\arguments{
\item{x}{Data matrix 'x'}

\item{y}{Response vector/two-column matrix 'y' (see: family); number of rows equal to nrow(x)}

\item{k}{Integer (0/1) kit indicator matrix; number of columns equal to ncol(x)}

\item{w}{Kit cost weight vector w of length nrow(k)}

\item{family}{Model family, should be one of: 'cox', 'mse'/'gaussian', or 'logistic, Default: 'gaussian'}

\item{print}{Level of verbosity in Fortran (may not be visible on all terminals); should be an integer between {range, range}, Default: 3}

\item{start}{Starting point generation method, see vignettes for details; should be an integer between {range,range}, Default: 2}

\item{verb}{Integer with additional integer values giving verbal feedback, Default: 1}

\item{kmax}{Maximum k step tested, by default all k are tested from k to maximum dimensionality}
}
\value{
Fitted casso-object
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
  data(ex)
  fit <- casso(x=ex_X, y=ex_Y, k=ex_K, w=ex_c, family='cox')
  fit
 }
}
}
\seealso{
\code{\link[survival]{coxph}},\code{\link[survival]{coxph.control}}
 \code{\link[stats]{glm}}
 \code{\link[casso]{character(0)}}
}
