% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\docType{methods}
\name{cost,oscar-method}
\alias{cost,oscar-method}
\title{Return total cost of model fit based on provided kit/variable costs vector}
\usage{
\S4method{cost}{oscar}(object, k)
}
\arguments{
\item{object}{Fit oscar S4-object}

\item{k}{Cardinality 'k' to compute total feature cost at}
}
\description{
Return total cost of model fit based on provided kit/variable costs vector
}
