% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\docType{methods}
\name{cost}
\alias{cost}
\alias{cost,oscar-methods}
\alias{cost,oscar-method}
\title{Return total cost of model fit based on provided kit/variable costs vector}
\usage{
cost(object, k)

\S4method{cost}{oscar}(object, k)
}
\arguments{
\item{object}{Fit oscar S4-object}

\item{k}{Cardinality 'k' to compute total feature cost at}
}
\value{
Numeric value of total feature/kit cost at cardinality 'k'

Numeric value of total feature/kit cost at cardinality 'k'
}
\description{
Return total cost of model fit based on provided kit/variable costs vector

Return total cost of model fit based on provided kit/variable costs vector
}
