% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cv.casso}
\alias{cv.casso}
\title{Cross-validation for casso-fitted model objects over k-range}
\usage{
cv.casso(fit, fold = 10, seed = NULL, verb = 0, ...)
}
\arguments{
\item{fit}{PARAM_DESCRIPTION}

\item{fold}{PARAM_DESCRIPTION, Default: 10}

\item{seed}{PARAM_DESCRIPTION, Default: NULL}

\item{verb}{PARAM_DESCRIPTION, Default: 0}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[casso]{character(0)}}
 \code{\link[stats]{predict.glm}}
 \code{\link[survival]{predict.coxph}},\code{\link[survival]{coxph}},\code{\link[survival]{Surv}}
}
