% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{cv.visu}
\alias{cv.visu}
\title{Visualize cross-validation as a function of k}
\usage{
cv.visu(
  cvs,
  add = FALSE,
  main = "OSCAR cross-validation",
  xlab = "Cardinality 'k'",
  ylab = "CV performance",
  ...
)
}
\arguments{
\item{cvs}{PARAM_DESCRIPTION}

\item{add}{PARAM_DESCRIPTION}

\item{main}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
