% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.glm.control.mod}
\alias{.glm.control.mod}
\title{Modified glm.control allowing 0 in maxit}
\usage{
.glm.control.mod(epsilon = 1e-08, maxit = 25, trace = FALSE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{weights}{PARAM_DESCRIPTION, Default: rep.int(1, nobs)}

\item{start}{PARAM_DESCRIPTION, Default: NULL}

\item{etastart}{PARAM_DESCRIPTION, Default: NULL}

\item{mustart}{PARAM_DESCRIPTION, Default: NULL}

\item{offset}{PARAM_DESCRIPTION, Default: rep.int(0, nobs)}

\item{family}{PARAM_DESCRIPTION, Default: gaussian()}

\item{control}{PARAM_DESCRIPTION, Default: list()}

\item{intercept}{PARAM_DESCRIPTION, Default: TRUE}

\item{singular.ok}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS TODO
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
