% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.glm.fit.mod}
\alias{.glm.fit.mod}
\title{Modified function from stats::glm.fit, allowing 0 maxit in control}
\usage{
.glm.fit.mod(
  x,
  y,
  weights = rep.int(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep.int(0, nobs),
  family = gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION}

\item{weights}{PARAM_DESCRIPTION, Default: rep.int(1, nobs)}

\item{start}{PARAM_DESCRIPTION, Default: NULL}

\item{etastart}{PARAM_DESCRIPTION, Default: NULL}

\item{mustart}{PARAM_DESCRIPTION, Default: NULL}

\item{offset}{PARAM_DESCRIPTION, Default: rep.int(0, nobs)}

\item{family}{PARAM_DESCRIPTION, Default: gaussian()}

\item{control}{PARAM_DESCRIPTION, Default: list()}

\item{intercept}{PARAM_DESCRIPTION, Default: TRUE}

\item{singular.ok}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS TODO
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
