% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\docType{methods}
\name{feat,oscar-method}
\alias{feat,oscar-method}
\title{Return named vector of feature indices with a given k that are non-zero}
\usage{
\S4method{feat}{oscar}(object, k)
}
\arguments{
\item{object}{Fit oscar S4-object}

\item{k}{Cardinality 'k' to extract non-zero features at}
}
\description{
Return named vector of feature indices with a given k that are non-zero
}
