% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\name{feat}
\alias{feat}
\title{Return named vector of feature indices with a given k that are non-zero}
\usage{
feat(object, k)
}
\description{
Return named vector of feature indices with a given k that are non-zero
}
