% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\docType{methods}
\name{kits,oscar-method}
\alias{kits,oscar-method}
\title{Return named vector of indices for kits with a given k that are non-zero}
\usage{
\S4method{kits}{oscar}(object, k)
}
\arguments{
\item{object}{Fit oscar S4-object}

\item{k}{Cardinality 'k' to extract kit indices at}
}
\description{
Return named vector of indices for kits with a given k that are non-zero
}
