% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.R
\docType{package}
\name{oscar-package}
\alias{oscar-package}
\title{oscar: Optimal Subset Cardinality Regression}
\description{
OSCAR models utilize the L0-pseudonorm to select an optimal subset of features that generalizes linear regression models to a variety of families. Currently supported models include conventional Gaussian regression (family="mse" or family="gaussian"), Binomial/Logistic regression (family="logistic"), and Cox proportional hazards modeling (family="cox").
}
\references{
Halkola AS, Joki K, Mirtti T, Mäkelä MM, Aittokallio T, Laajala TD (2023) OSCAR: Optimal subset cardinality regression using the L0-pseudonorm with applications to prognostic modelling of prostate cancer. PLoS Comput Biol 19(3): e1010333. \doi{10.1371/journal.pcbi.1010333}
}
