% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{oscar.binplot}
\alias{oscar.binplot}
\title{Visualize binary indicator matrix optionally coupled with cross-validation performance}
\usage{
oscar.binplot(
  fit,
  cv,
  kmax,
  collines = TRUE,
  rowlines = TRUE,
  cex.axis = 0.6,
  heights = c(0.2, 0.8),
  ...
)
}
\arguments{
\item{fit}{TODO}

\item{cv}{TODO}

\item{kmax}{TODO}

\item{collines}{TODO}

\item{rowlines}{TODO}

\item{cex.axis}{TODO}

\item{heights}{TODO}

\item{...}{Additional parameters passed on to hamlet::hmap}
}
\description{
TODO
}
