% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitS4.R
\name{oscar.control}
\alias{oscar.control}
\title{Control OSCAR optimizer parameters}
\usage{
oscar.control(
  x,
  family,
  start = 2,
  in_mrounds = 5000,
  in_mit = 5000,
  in_mrounds_esc = 5000,
  in_b1,
  in_b2 = 3,
  in_b,
  in_m = 0.01,
  in_m_clarke = 0.01,
  in_c = 0.1,
  in_r_dec,
  in_r_inc = 10^5,
  in_eps1 = 5 * 10^(-5),
  in_eps,
  in_crit_tol = 10^(-5),
  na = 4,
  mcu = 7,
  mcinit = 7,
  tolf = 10^(-5),
  tolf2 = 10^4,
  tolg = 10^(-5),
  tolg2 = tolg,
  eta = 0.5,
  epsL = 0.125
)
}
\arguments{
\item{start}{Starting point generation method, see vignettes for details; should be an integer between {range,range}, Default: 2}

\item{in_mrounds}{DESCRIPTION}

\item{in_mit}{DESCRIPTION}

\item{in_mrounds_esc}{DESCRIPTION}

\item{in_b1}{DESCRIPTION}

\item{in_b2}{DESCRIPTION}

\item{in_m}{DESCRIPTION}

\item{in_m_clarke}{DESCRIPTION}

\item{in_c}{DESCRIPTION}

\item{in_r_dec}{DESCRIPTION}

\item{in_r_inc}{DESCRIPTION}

\item{in_eps1}{DESCRIPTION}

\item{in_eps}{DESCRIPTION}

\item{in_crit_tol}{DESCRIPTION}

\item{na}{DESCRIPTION}

\item{mcu}{DESCRIPTION}

\item{mcinit}{DESCRIPTION}

\item{tolf}{DESCRIPTION}

\item{tolf2}{DESCRIPTION}

\item{tolg}{DESCRIPTION}

\item{tolg2}{DESCRIPTION}

\item{eta}{DESCRIPTION}

\item{epsL}{DESCRIPTION}
}
\value{
A list of parameters for the OSCAR optimizers.
}
\description{
Fine-tuning the parameters available for the DBDC and LMBM optimizers. See oscar documentation for the optimization algorithms for further details.
}
\details{
This function sanity checks and provides reasonable DBDC and LMBM optimization tuning parameters. User may override custom values, though sanity checks will prevent unreasonable values and replace them.
}
\examples{
\dontrun{
if(interactive()){
  oscar.control() # Return a list of default parameters
 }
}
}
