% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{oscar.cost.after}
\alias{oscar.cost.after}
\title{Return total cost of model fits if the cost is not included in the oscar object
If at least one measurement from a kit is included in the model, the kit cost is added.}
\usage{
oscar.cost.after(object)
}
\arguments{
\item{object}{Fit oscar S4-object}
}
\description{
Return total cost of model fits if the cost is not included in the oscar object
If at least one measurement from a kit is included in the model, the kit cost is added.
}
