% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{pareto}
\alias{pareto}
\title{Visualize pareto-optimal front in respect to cross-validation and variable/kit costs}
\usage{
pareto(object, cv, k = 1:object@kmax, plot = TRUE)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{cv}{PARAM_DESCRIPTION}

\item{k}{PARAM_DESCRIPTION}

\item{plot}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

}
