% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\name{plot,casso-method}
\alias{plot,casso-method}
\title{Plot casso-coefficients as a function of k and override default plot generic}
\usage{
\S4method{plot}{casso}(x, y, k = 1:x@kmax, add = FALSE, intercept = FALSE, ...)
}
\description{
Plot casso-coefficients as a function of k and override default plot generic
}
