% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{plot.pareto.costCI}
\alias{plot.pareto.costCI}
\alias{plot.smoothing.spline}
\title{Visualize cost and C-index with paretofront without oscar object}
\usage{
plot.pareto.costCI(
  costs,
  cindexes,
  nzeros,
  col,
  ncolors,
  ylab = "cindex",
  xlab = "cost",
  pch = 19,
  bg = col,
  cex = 0.2,
  ylim = c(min(cindexes), max(cindexes)),
  xlim = c(min(costs), max(costs)),
  las = 1,
  title = "",
  width.of.layout = c(2.2, 0.8),
  height.of.layout = c(1, 1),
  mar.1st = c(4, 4, 2, 0),
  mar.2nd = c(2.5, 0, 2, 2),
  monochrome = F,
  cex.pareto = 0.5,
  col.paretoline = "grey"
)

plot.smoothing.spline(
  x,
  y,
  ylab = "y",
  xlab = "x",
  a.title = "",
  df = ceiling(length(unique(x)))/2
)
}
\arguments{
\item{costs}{Vector including costs values}

\item{cindexes}{Vector including C-index values}

\item{nzeros}{Vector including the number of non-zeros of the model in each point}

\item{col}{Vector including desired colors for each number of non-zeros, Default : 'black'}

\item{ncolors}{Number of required colors, Default: max(nzeros)-min(nzeros)+1}

\item{ylab}{Title for y-axis}

\item{xlab}{Title for x-axis}

\item{mar.1st}{Marginals for the main plot, Default: c(4,4,2,0)}

\item{mar.2nd}{Marginals for the legend, Default: c(2.5,0,2,2)}

\item{monochrome}{If the plot should be with colors (F) or monochrome (T), Default: F}

\item{x}{Values for x-axis}

\item{y}{Values for y-axis}

\item{a.title}{Main title for the scatter plot with spline}

\item{df}{Degrees of freedom, Default: 'ceiling(length(unique(x)))/2)'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
