% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{visu}
\alias{visu}
\title{Target function value and total kit cost as a function of number of kits included}
\usage{
visu(
  object,
  y = c("target", "cost", "goodness", "cv"),
  cols = c("red", "blue"),
  legend = "top",
  mtexts = TRUE,
  ...
)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{y}{PARAM_DESCRIPTION, Default: c("target", "cost", "goodness", "cv")}

\item{cols}{PARAM_DESCRIPTION, Default: c("red", "blue")}

\item{legend}{PARAM_DESCRIPTION, Default: 'top'}

\item{mtexts}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
