% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/visu.R
\name{bs.oscar}
\alias{bs.oscar}
\alias{bs.visu}
\title{Bootstrapping for oscar-fitted model objects}
\usage{
bs.oscar(fit, bootstrap = 100, seed = NULL, verb = 0)

bs.visu(bs, intercept = FALSE, add = FALSE)
}
\arguments{
\item{fit}{oscar-model object}

\item{bootstrap}{Number of bootstrapped datasets, Default: 100}

\item{seed}{Random seed for reproducibility with NULL indicating that it is not set, Default: NULL}

\item{verb}{Level of verbosity with higher integer giving more information, Default: 0}

\item{bs}{PARAM_DESCRIPTION}

\item{intercept}{PARAM_DESCRIPTION, Default: FALSE}

\item{add}{PARAM_DESCRIPTION}
}
\value{
3-dimensional array with dimensions corresponding to k-steps, beta coefficients, and bootstrap runs

OUTPUT_DESCRIPTION
}
\description{
This model bootstraps the fitting of a given oscar object (re-fits the model for data that is equal in size but sampled with replacement). The output objects give insight into robustness of the oscar-coefficient path, as well as relative importance of model objects.

FUNCTION_DESCRIPTION
}
\details{
TODO

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
