% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/visu.R
\name{cv.oscar}
\alias{cv.oscar}
\alias{cv.visu}
\title{Cross-validation for oscar-fitted model objects over k-range}
\usage{
cv.oscar(fit, fold = 10, seed = NULL, verb = 0)

cv.visu(
  cvs,
  add = FALSE,
  main = "OSCAR cross-validation",
  xlab = "Cardinality 'k'",
  ylab = "CV performance",
  ...
)
}
\arguments{
\item{fit}{oscar-model object}

\item{fold}{Number of cross-validation folds, Default: 10}

\item{seed}{Random seed for reproducibility with NULL indicating that it is not set, Default: NULL}

\item{verb}{Level of verbosity with higher integer giving more information, Default: 0}

\item{cvs}{PARAM_DESCRIPTION}

\item{add}{PARAM_DESCRIPTION}

\item{main}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
A matrix with goodness of fit over folds and k-values

OUTPUT_DESCRIPTION
}
\description{
Create a cross-validation matrix with the chosen goodness metric with n-folds. Based on the goodness metric, one ought to pick optimal cardinality (parameter 'k').

FUNCTION_DESCRIPTION
}
\details{
TODO

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}

\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
