% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.glm.fit.mod}
\alias{.glm.fit.mod}
\title{Modified function from stats::glm.fit, allowing 0 maxit in control}
\usage{
.glm.fit.mod(
  x,
  y,
  weights = rep.int(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep.int(0, nobs),
  family = gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\description{
Modified function from stats::glm.fit, allowing 0 maxit in control
}
