% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\name{show,casso-method}
\alias{show,casso-method}
\alias{coef,casso-method}
\alias{plot,casso-method}
\alias{feat}
\alias{kits}
\title{Showing casso-objects}
\usage{
\S4method{show}{casso}(object)

\S4method{coef}{casso}(object, k)

\S4method{plot}{casso}(x, y, k = 1:x@kmax, add = FALSE, intercept = FALSE, ...)

feat(object, k)

kits(object, k)
}
\description{
Showing casso-objects

Extract coefficients of casso-objects

Plot casso-coefficients as a function of k and override default plot generic

Return named vector of feature indices with a given k that are non-zero

Return named vector of kits indices with a given k that are non-zero
}
