% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\name{show,oscar-method}
\alias{show,oscar-method}
\alias{coef,oscar-method}
\alias{predict,oscar-method}
\alias{plot,oscar-method}
\alias{feat}
\alias{kits}
\alias{cost}
\alias{cost.after}
\title{Showing oscar-objects}
\usage{
\S4method{show}{oscar}(object)

\S4method{coef}{oscar}(object, k)

\S4method{predict}{oscar}(
  object,
  k,
  type = c("response", "link", "nonzero", "coefficients", "label"),
  newdata = object@x
)

\S4method{plot}{oscar}(x, y, k = 1:x@kmax, add = FALSE, intercept = FALSE, ...)

feat(object, k)

kits(object, k)

cost(object, k)

cost.after(object, kit.matrix, cost.vector)
}
\description{
Showing oscar-objects

Extract coefficients of oscar-objects

Predicting based on oscar-objects

Plot oscar-coefficients as a function of k and override default plot generic

Return named vector of feature indices with a given k that are non-zero

Return named vector of kits indices with a given k that are non-zero

Return total cost of model fit based on provided kit/variable costs vector

Return total cost of model fits if the cost is not included in the oscar object
If at least one measurement from a kit is included in the model, the kit cost is added.
}
