% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\name{show,oscar-method}
\alias{show,oscar-method}
\title{Showing oscar-objects}
\usage{
\S4method{show}{oscar}(object)
}
\arguments{
\item{object}{Fit oscar S4-object}
}
\value{
Outputs raw text describing key characteristics of the oscar-object
}
\description{
Showing oscar-objects
}
