% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcsS4.R
\name{kits}
\alias{kits}
\title{Return named vector of kits indices with a given k that are non-zero}
\usage{
kits(object, k)
}
\description{
Return named vector of kits indices with a given k that are non-zero
}
