% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pckg.R
\docType{package}
\name{oscar-package}
\alias{oscar-package}
\title{oscar: Optimal Subset Cardinality Regression}
\description{
OSCAR models utilize the L0-pseudonorm to select an optimal subset of features that generalizes linear regression models to a variety of families. Currently supported models include conventional Gaussian regression (family="mse" or family="gaussian"), Binomial/Logistic regression (family="logistic"), and Cox proportional hazards modeling (family="cox").
}
\references{
Halkola A.S., Joki K., Mirtti T., Makela M.M., Aittokallio T., Laajala, T.D. OSCAR: Optimal subset cardinality regression using the L0-pseudonorm. To be submitted.
}
