% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{oscar.bs.plot}
\alias{oscar.bs.plot}
\title{Bootstrap + cross-validation heatmap plot}
\usage{
oscar.bs.plot(
  fit,
  bs,
  kmax,
  cex.axis = 0.6,
  palet = colorRampPalette(c("orange", "red", "black", "blue", "cyan"))(dim(bs)[3]),
  nbins = 100,
  Colv = NA,
  Rowv = NA,
  ...
)
}
\arguments{
\item{fit}{TODO}

\item{bs}{TODO}

\item{kmax}{TODO}

\item{cex.axis}{TODO}

\item{palet}{TODO}

\item{nbins}{TODO}

\item{Colv}{TODO}

\item{Rowv}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
