% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{oscar.bs.visu}
\alias{oscar.bs.visu}
\title{Visualize bootstrapping of a fit oscar object}
\usage{
oscar.bs.visu(bs, intercept = FALSE, add = FALSE)
}
\arguments{
\item{bs}{Bootstrapped 3-dimensional array for an oscar object as produced by oscar.bs}

\item{intercept}{Whether model intercept should be plotted also as a coefficient, Default: FALSE}

\item{add}{Should plot be added on top of an existing plot device}
}
\description{
This function visualizes bootstrapped model coefficients over multiple bootstrap runs as lines in a graph
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
