% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{oscar.cost.after}
\alias{oscar.cost.after}
\title{Return total cost of model fits if the cost is not included in the oscar object}
\usage{
oscar.cost.after(object)
}
\arguments{
\item{object}{Fit oscar S4-object}
}
\value{
A vector for numeric values of total kit costs at different cardinalities.
}
\description{
If at least one measurement from a kit is included in the model, the kit cost is added.
}
\examples{
if(interactive()){
  data(ex)
  fit <- oscar(x=ex_X, y=ex_Y, k=ex_K, w=ex_c, family='cox')
  oscar.cost.after(fit)
}

}
