% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitS4.R, R/pckg.R
\docType{package}
\name{oscar}
\alias{oscar}
\title{FUNCTION_TITLE}
\usage{
oscar(
  x,
  y,
  k,
  w,
  family = "cox",
  print = 3,
  start = 2,
  verb = 1,
  kmax,
  in_mrounds = 5000,
  in_mit = 5000,
  in_mrounds_esc = 5000,
  in_b1,
  in_b2 = 3,
  in_b,
  in_m = 0.2,
  in_m_clarke = 0.01,
  in_c = 0.1,
  in_r_dec,
  in_r_inc = 10^7,
  in_eps1 = 5 * 10^(-5),
  in_eps,
  in_crit_tol
)
}
\arguments{
\item{x}{Data matrix 'x'}

\item{y}{Response vector/two-column matrix 'y' (see: family); number of rows equal to nrow(x)}

\item{k}{Integer (0/1) kit indicator matrix; number of columns equal to ncol(x)}

\item{w}{Kit cost weight vector w of length nrow(k)}

\item{family}{Model family, should be one of: 'cox', 'mse'/'gaussian', or 'logistic, Default: 'gaussian'}

\item{print}{Level of verbosity in Fortran (may not be visible on all terminals); should be an integer between {range, range}, Default: 3}

\item{start}{Starting point generation method, see vignettes for details; should be an integer between {range,range}, Default: 2}

\item{verb}{Integer with additional integer values giving verbal feedback, Default: 1}

\item{kmax}{Maximum k step tested, by default all k are tested from k to maximum dimensionality}
}
\value{
Fitted oscar-object
}
\description{
FUNCTION_DESCRIPTION

OSCAR models utilize the L0-pseudonorm to select an optimal subset of features that generalizes linear regression models to a variety of families. Currently supported models include conventional Gaussian regression (family="mse" or family="gaussian"), Binomial/Logistic regression (family="logistic"), and Cox proportional hazards modeling (family="cox").
}
\details{
DETAILS
}
\section{oscar functionality, list of functions}{
NA
}

\examples{
\dontrun{
if(interactive()){
  data(ex)
  fit <- oscar(x=ex_X, y=ex_Y, k=ex_K, w=ex_c, family='cox')
  fit
 }
}
}
\seealso{
\code{\link[survival]{coxph}},\code{\link[survival]{coxph.control}}
 \code{\link[stats]{glm}}
}
