% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitS4.R, R/pckg.R
\docType{package}
\name{oscar}
\alias{oscar}
\title{Main OSCAR fitting function}
\usage{
oscar(
  x,
  y,
  k,
  w,
  family = "cox",
  metric,
  solver = 1,
  verb = 1,
  print = 3,
  kmax,
  sanitize = TRUE,
  control,
  ...
)
}
\arguments{
\item{x}{Data matrix 'x'}

\item{y}{Response vector/two-column matrix 'y' (see: family); number of rows equal to nrow(x)}

\item{k}{Integer (0/1) kit indicator matrix; number of columns equal to ncol(x), Default: Unit diagonal indicator matrix}

\item{w}{Kit cost weight vector w of length nrow(k), Default: Equal cost for all variables}

\item{family}{Model family, should be one of: 'cox', 'mse'/'gaussian', or 'logistic, Default: 'cox'}

\item{metric}{Goodness metric, Default(s): Concordance index for Cox, MSE for Gaussian, and AUC for logistic regression}

\item{solver}{Solver used in the optimization, should be  1/'DBDC' or 2/'LMBM', Default: 1.}

\item{verb}{Level of verbosity in R, Default: 1}

\item{print}{Level of verbosity in Fortran (may not be visible on all terminals); should be an integer between {range, range}, Default: 3}

\item{kmax}{Maximum k step tested, by default all k are tested from k to maximum dimensionality, Default: ncol(x)}

\item{sanitize}{Whether input column names should be cleaned of potentially problematic symbols, Default: TRUE}

\item{control}{Tuning parameters for the optimizers, see function oscar.control(), Default: see ?oscar.control}

\item{...}{Additional parameters}
}
\value{
Fitted oscar-object
}
\description{
This function fits an OSCAR model object to the provided training data with the desired model family.

OSCAR models utilize the L0-pseudonorm to select an optimal subset of features that generalizes linear regression models to a variety of families. Currently supported models include conventional Gaussian regression (family="mse" or family="gaussian"), Binomial/Logistic regression (family="logistic"), and Cox proportional hazards modeling (family="cox").
}
\details{
DETAILS
}
\section{oscar functionality, list of functions}{
NA
}

\examples{
\dontrun{
if(interactive()){
  data(ex)
  fit <- oscar(x=ex_X, y=ex_Y, k=ex_K, w=ex_c, family='cox')
  fit
 }
}
}
